+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 5.4.0 [commit logs](https://github.com/biplane/yandex-direct/compare/5.3.0...5.4.0)
+
 ### Changed
 
 * Обновлены сервисы для совмести с последней версией API 5, включая изменения от 2022-04-29.